# API-SPEC.yaml
openapi: 3.0.3
info:
  title: SFOps API
  description: Salesforce Deployment SaaS Platform API
  version: 1.0.0
  contact:
    name: SFOps Support
    email: support@sfops.io
servers:
  - url: https://api.sfops.io/v1
    description: Production
  - url: https://staging-api.sfops.io/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - bearerAuth: []
...
paths:
...
  /tenants/{tenantId}/orgs/{orgId}/snapshots:
    get:
      tags: [Snapshots]
      summary: List snapshots for an org
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/orgId'
      responses:
        '200':
          description: List of snapshots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Snapshot'

    post:
      tags: [Snapshots]
      summary: Create new snapshot
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/orgId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                metadataTypes:
                  type: array
                  items:
                    type: string
      responses:
        '202':
          description: Snapshot job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
  /tenants/{tenantId}/snapshots/{snapshotId}:
    get:
      tags: [Snapshots]
      summary: Get snapshot details
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/snapshotId'
      responses:
        '200':
          description: Snapshot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'

  # Diffs
...
  /tenants/{tenantId}/packages:
    get:
      tags: [Packages]
      summary: List deployment packages
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: List of packages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Package'

    post:
      tags: [Packages]
      summary: Create deployment package
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, diffId, selectedComponents]
              properties:
                name:
                  type: string
                description:
                  type: string
                diffId:
                  type: string
                selectedComponents:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Package created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
  /tenants/{tenantId}/packages/{packageId}:
    get:
      tags: [Packages]
      summary: Get package details
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/packageId'
      responses:
        '200':
          description: Package details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'

  # Deployments
...
  /tenants/{tenantId}/pipelines:
    get:
      tags: [Pipelines]
      summary: List pipelines
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: List of pipelines
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pipeline'

    post:
      tags: [Pipelines]
      summary: Create pipeline
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineCreate'
      responses:
        '201':
          description: Pipeline created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
  /tenants/{tenantId}/pipelines/{pipelineId}:
    get:
      tags: [Pipelines]
      summary: Get pipeline details
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/pipelineId'
      responses:
        '200':
          description: Pipeline details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
  /tenants/{tenantId}/pipelines/{pipelineId}/runs:
    get:
      tags: [Pipelines]
      summary: List pipeline runs
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/pipelineId'
      responses:
        '200':
          description: List of pipeline runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineRun'

  # Git Integration
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
    snapshotId:
      name: snapshotId
      in: path
      required: true
      schema:
        type: string
...
    packageId:
      name: packageId
      in: path
      required: true
      schema:
        type: string
...
    pipelineId:
      name: pipelineId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Snapshot:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        componentCount:
          type: integer
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Package:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        components:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    Pipeline:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        stages:
          type: array
          items:
            $ref: '#/components/schemas/PipelineStage'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/PipelineTrigger'
    PipelineCreate:
      type: object
      required: [name, stages]
      properties:
        name:
          type: string
        description:
          type: string
        stages:
          type: array
          items:
            $ref: '#/components/schemas/PipelineStage'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/PipelineTrigger'
    PipelineStage:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [validate, test, deploy, approval]
        targetOrgId:
          type: string
        requiresApproval:
          type: boolean
    PipelineTrigger:
      type: object
      properties:
        type:
          type: string
          enum: [git_push, pull_request, schedule, manual]
        config:
          type: object
    PipelineRun:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed]
        trigger:
          type: string
        startedAt:
          type: string
          format: date-time
...